feat: v2.7.5 NUT compatibility hardening

+++ src/drivers/hidparser.h
@@ -68,7 +68,7 @@ int FindObject(HIDDesc_t *pDesc_arg, HIDData_t *pData);
 
 HIDData_t *FindObject_with_Path(HIDDesc_t *pDesc_arg, HIDPath_t *Path, uint8_t Type);
 
-HIDData_t *FindObject_with_ID(HIDDesc_t *pDesc_arg, uint8_t ReportID, uint8_t Offset, uint8_t Type);
+HIDData_t *FindObject_with_ID(HIDDesc_t *pDesc_arg, uint8_t ReportID, uint16_t Offset, uint8_t Type);
 
 HIDData_t *FindObject_with_ID_Node(HIDDesc_t *pDesc_arg, uint8_t ReportID, HIDNode_t Node);
 /*
